backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
google-genai==1.45.0
python-dotenv==1.1.0
